requirements.txt
Flask
pymongo
Flask-PyMongo
requests>=2.4.1
pycryptodome
python-jose
